% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNPsearch.R
\name{SNPsearch}
\alias{SNPsearch}
\title{SNPsearch}
\usage{
SNPsearch(rsid)
}
\arguments{
\item{rsid}{rsid of SNP}
}
\value{
frame of mutations asssociated with rsid
}
\description{
SNPsearch
}
\examples{
SNPsearch("rs333")
}
