% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNPshiftStop.R
\name{SNPshiftStop}
\alias{SNPshiftStop}
\title{SNPshiftStop}
\usage{
SNPshiftStop(
  gene.symbol,
  taxon = "human",
  mutation.from.start = NA,
  ref.bp = "-",
  alt.bp = "-",
  ref.aa = NA,
  alt.aa = NA,
  aa.mut.pos = NA,
  start.loss = FALSE,
  ncbi.api.key = NULL
)
}
\arguments{
\item{gene.symbol}{Gene symbol with frameshift mutation}

\item{taxon}{provide taxon character string or number; found using NCBI (default human (9606))}

\item{ref.bp}{reference base pair (default is '"-"')}

\item{alt.bp}{alternate base pair (default is '"-"')}

\item{ref.aa}{reference amino acid (default is NA)}

\item{alt.aa}{alternate amino acid (default is NA)}

\item{aa.mut.pos}{amino acid mutation position (default is NA)}

\item{start.loss}{does this mutation contain a start.loss (default is TRUE)}

\item{mutation}{Location of mutation relative to start or}
}
\value{
(gene.symbol, aa.seq, mutated.aa.seq, shortened.of.total.aa, dna.seq, mutated.dna.seq)
}
\description{
SNPshiftStop
}
\examples{
SNPshiftStop(gene.symbol="CFTR", taxon="human", mutation=1213,
ref.bp="T", ref.aa="F", alt.aa="L", aa.mut.pos=405)
}
